main landed #4343, which edits the same list_runs contract, handler and Fleet board. Both opt-in counts stay: count answers the pager's total, and countLive answers the Live runs tile. #4343's changes to the row mappers (the live-status constants, tachoPlace, place on a ledger row, and the session's directory columns) are ported into lib/run-item.ts, where the batch moved that section. A stale run's light wins over its parked or paused word in RowStatusBadge, as #4343 set for the status cell.

The run-rows test that counted Live runs from the rows now reads the workspace count, which counts a paused run as open, as the live chip lists it. atlas.sum, the storage manifest, the capability schema docs and the message types are regenerated. main's newest migration (20260925235000) precedes the batch's, so it keeps its stamp.

+++ packages/handlers/src/lib/run-list-work.ts
@@ -201,11 +201,14 @@ export const postgresRunGitDiffs: ReadRunGitDiffs = async (
       .select({
         chain: sessions.sessionUuid,
         root: sessions.rootSessionUuid,
-        added: sql<number>`coalesce(sum(${files.linesAdded}), 0)::int`.mapWith(
-          Number,
-        ),
+        // bigint: one file's count can pass 2^31 - 1 (#3944, S-02), and an
+        // ::int cast then fails the page's whole read.
+        added:
+          sql<number>`coalesce(sum(${files.linesAdded}), 0)::bigint`.mapWith(
+            Number,
+          ),
         removed:
-          sql<number>`coalesce(sum(${files.linesRemoved}), 0)::int`.mapWith(
+          sql<number>`coalesce(sum(${files.linesRemoved}), 0)::bigint`.mapWith(
             Number,
           ),
       })
